lastCommit

+++ src/app/page-signup/page-signup.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { FormBuilder, Validators } from '@angular/forms';
 import { Iuser } from './user.model';
 import { ServiceUserService } from './service-user.service';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-page-signup',
@@ -19,7 +20,10 @@ export class PageSignupComponent implements OnInit {
     pseudo : ['',Validators.required],
   })
 
-  constructor(private formBuilder : FormBuilder,private serviceUser:ServiceUserService){}
+  constructor(
+    private formBuilder : FormBuilder,private serviceUser:ServiceUserService,
+    private router:Router 
+    ){}
   ngOnInit(): void {
    const localData = localStorage.getItem('signUpUsers');
    if(localData != null){
@@ -52,6 +56,25 @@ export class PageSignupComponent implements OnInit {
 };
 
 onSignUp(){
+  let signUpUsers: any[] = JSON.parse(localStorage.getItem("signUpUsers") as string);
+  const pseudoExist = signUpUsers.find(m=> m.pseudo == this.signupObj.pseudo && m.pseudo == this.signupObj.pseudo);
+  const emailExist = signUpUsers.find(m=> m.email == this.signupObj.email && m.email ==this.signupObj.email);
+ 
+  if(pseudoExist){
+    const sonPseudo = new Audio;
+      sonPseudo.src="../../assets/audio/pseudoUsed.mp3";
+      sonPseudo.play();
+    alert("le pseudo existe déja ❌");
+    return
+  }
+
+  if(emailExist){
+    const sonEmail = new Audio;
+      sonEmail.src="../../assets/audio/usedEmail.mp3";
+      sonEmail.play();
+    alert("l'adresse email existe déja ❌ ")
+    return
+  }
   this.signUpUsers.push(this.signupObj);
   localStorage.setItem("signUpUsers", JSON.stringify(this.signUpUsers));
   sessionStorage.setItem("currentUser", this.signupObj.pseudo)
@@ -64,7 +87,8 @@ onSignUp(){
     pseudo: ''
 
 };
+window.location.href="http://localhost:4200/"
+  this.router.navigate(["/"]) 
+  window.location.href="http://localhost:4200/"
 }
-
-
 }
